distrib_systems_tasks/weeks/week-16/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
bandit>=1.7.0
pip-audit>=2.5.0